containers/extract_upload_to_gcs/scripts/requirements.txt
pandas
mysql-connector-python
google-cloud-storage
